utils cachedFetch method
